+++ structures/scaletor/tunnings/eight/standard.js
@@ -1,0 +1,3 @@
+const standard = [8, 3, 11, 6, 1, 8, 3, 10];
+
+export default standard;