+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 This file documents all notable changes made up to version 2.0.0 for each project.
 
 ## Unreleased
+- Add ABASI 7-8 string templates and tunnings for 6-7-8 strings.
 
 ## [2.0.2] - 2024-10-05
 ### Changes